Implement prepend instrumentation for Bunny (#599)
* default prepend to true

* add prepend for bunny

* fox for bunny prepend

* changed bunny prepend to module

* updated bunny testing for chain and prepend

* updated bunny

* update bunny instrumentation

* fix require_relative

* reset default for config not related to bunny

* added config check for bunny

* DRYed up bunny instrumentation

* fix leftover end

* fix missing args

* cleaned up and simplified
